Add nix-optimise as a systemd service
The nix-optimise systemd service will run the command
    `nix-store --optimise'

I have configured this to run on a weekly basis.
It goes through the nix store, finds duplicate packages, and hardlinks
the duplicates together, and removing any duplicates.
This can drastically reduce the amount of disk space needed to keep
the store in a usable position, and it does not affect the way the
store works in any way.

However, because this is such an expensive operation, I configured it
only happen once a week.