Remove DisplayLink USB Hub's ethernet port from config
If I were to keep this in the configuration, any system that used this
portion of the hardware configuration would wait to boot until either
this systemd startup timed out, or the DisplayLink's ethernet port was
plugged in.

However, on most devices, this is not actually present, so I this lead
to very long boot times while waiting for this network interface to try
to come up.
So, I figured it out and removed it.