Add output from nvidia-settings GUI
This program was meant to write things out to a file in
/etc/X11/xorg.conf that would manage the X11 server for me. However,
because I'm on NixOS, this file and directory will be read-only.

So, I tried to override the default services.xserver.config using this
option, but it does not seem to be taking effect the way I want it to.
Instead, it is generating both this portion, and the portion that was
there before, and I think that is causing problems for the X server.