This config sets up both monitors with Xinerama
While this does allow BOTH monitors to run at the same time, the
problem here is that the monitors are seen "together", thus the
biggest a window can be on a single monitor is limited to the largest
resolution of the smallest monitor.

Thus, "full screen" is 1600x900, which doesn't really work.
Then, virtual full screen is split OVER the 2 screens, which is also
what I don't want.

I am switching to `xrandr` next and seeing if that improves the situation.