Remove rustc and cargo from system packages, leaving rustup
We need to make sure that the versions of rustc and cargo are
compatible, and the easiest way to do that is have rustup manage them
for us. Plus, rustup only affects the current user, so that means the
assurances NixOS provides (immutable store, etc.) are not affected.

Switch from rls to rust-analyzer, because the former is deprecated,
and rust-analyzer is becoming the main LSP server for the Rust language.